model/requirements.txt
flask-restful==0.3.8
redis==3.4.1
tensorflow-cpu==2.1.0
matplotlib==3.2.1
